refactor: migrate `getCurrentRange` to selection
